Aplicando correcciones a los comentarios
