+++ frontend/www/js/omegaup/components/qualitynomination/ReviewerPopup.vue
@@ -1,112 +1,110 @@
 <template>
-  <omegaup-popup
-    :reviewer-nomination="true"
-    :possible-tags="PROBLEM_CATEGORIES"
-    @submit="$emit('submit', tag, qualitySeal, selectedPublicTags)"
-  >
-    <template #link-title>
-      {{ T.reviewerNomination }}
-    </template>
-    <template #popup-content="slotProps">
-      <div class="title-text">
-        {{ T.reviewerNominationFormTitle }}
-      </div>
-      <div class="form-group">
-        <label class="control-label">
-          {{ T.reviewerNominationQuality }}
-        </label>
-        <br />
-        <label class="radio-inline"
-          ><input v-model="qualitySeal" type="radio" :value="true" />
-          {{ T.wordsYes }}</label
-        >
-        <label class="radio-inline"
-          ><input v-model="qualitySeal" type="radio" :value="false" />
-          {{ T.wordsNo }}</label
-        >
-      </div>
-      <div class="form-group">
-        <label class="control-label">
-          {{ T.reviewerNominationCategory }}
-          <ul class="tag-select">
-            <li
-              v-for="problemTopic in slotProps.sortedProblemTags"
-              :key="problemTopic.value"
-              class="tag-select"
-            >
-              <label class="tag-select"
-                ><input
-                  v-model="tag"
-                  type="radio"
-                  :value="problemTopic.value"
-                />
-                {{ problemTopic.text }}</label
-              >
-            </li>
-          </ul></label
-        >
-      </div>
-      <div class="form-group">
-        <vue-typeahead-bootstrap
-          :data="publicTags"
-          :serializer="publicTagsSerializer"
-          :placeholder="T.collecionOtherTags"
-          @hit="addOtherTag"
-        >
-        </vue-typeahead-bootstrap>
-        <br />
-        <div class="card-body table-responsive">
-          <table class="table table-striped">
-            <thead>
-              <th class="text-center" scope="col">
-                {{ T.contestEditTagName }}
-              </th>
-              <th class="text-center" scope="col">
-                {{ T.contestEditTagDelete }}
-              </th>
-            </thead>
-            <tbody>
-              <tr v-for="tag in publicTagsList" :key="tag">
-                <td>{{ getName(tag) }}</td>
-                <td class="text-center">
-                  <button
-                    type="button"
-                    class="btn btn-danger"
-                    :disabled="publicTagsList.length < 2"
-                    @click="removeTag(tag)"
+  <omegaup-overlay-popup @dismiss="onHide">
+    <transition name="fade">
+      <form data-reviewewr-popup class="h-auto w-auto" @submit.prevent="">
+        <div class="container-fluid d-flex align-items-start flex-column">
+          <template v-if="currentView === AvailableViews.Content">
+            <p class="h4 font-weight-bold pb-4 text-center w-100">
+              {{ T.reviewerNominationFormTitle }}
+            </p>
+            <div class="form-group w-100">
+              <label class="control-label">
+                {{ T.reviewerNominationQuality }}
+              </label>
+              <br />
+              <omegaup-radio-switch
+                :value.sync="qualitySeal"
+                :selected-value="qualitySeal"
+              ></omegaup-radio-switch>
+            </div>
+            <div class="form-group w-100">
+              <label class="control-label w-100">
+                {{ T.reviewerNominationCategory }}
+                <ul class="tag-select">
+                  <li
+                    v-for="problemTopic in sortedProblemTags"
+                    :key="problemTopic.value"
+                    class="tag-select"
                   >
-                    <font-awesome-icon :icon="['fas', 'trash']" />
-                  </button>
-                </td>
-              </tr>
-            </tbody>
-          </table>
+                    <label class="tag-label"
+                      ><input
+                        v-model="tag"
+                        type="radio"
+                        :value="problemTopic.value"
+                      />
+                      {{ problemTopic.text }}</label
+                    >
+                  </li>
+                </ul></label
+              >
+            </div>
+            <div class="form-group w-100">
+              <vue-typeahead-bootstrap
+                :data="publicTags"
+                :serializer="publicTagsSerializer"
+                :placeholder="T.collecionOtherTags"
+                @hit="addOtherTag"
+              >
+              </vue-typeahead-bootstrap>
+              <br />
+              <div class="card-body table-responsive w-100">
+                <table class="table table-striped w-100">
+                  <thead>
+                    <th class="text-center" scope="col">
+                      {{ T.contestEditTagName }}
+                    </th>
+                    <th class="text-center" scope="col">
+                      {{ T.contestEditTagDelete }}
+                    </th>
+                  </thead>
+                  <tbody>
+                    <tr v-for="tag in publicTagsList" :key="tag">
+                      <td>{{ getName(tag) }}</td>
+                      <td class="text-center">
+                        <button
+                          type="button"
+                          class="btn btn-danger"
+                          :disabled="publicTagsList.length < 2"
+                          @click="removeTag(tag)"
+                        >
+                          <font-awesome-icon :icon="['fas', 'trash']" />
+                        </button>
+                      </td>
+                    </tr>
+                  </tbody>
+                </table>
+              </div>
+            </div>
+            <div class="text-right">
+              <button
+                class="btn btn-primary mr-3"
+                type="submit"
+                :disabled="qualitySeal && !tag"
+                @click="onSubmit"
+              >
+                {{ T.wordsSend }}
+              </button>
+              <button class="btn btn-secondary" type="button" @click="onHide">
+                {{ T.wordsCancel }}
+              </button>
+            </div>
+          </template>
+          <template v-if="currentView === AvailableViews.Thanks">
+            <div class="w-100 h-100 h3 text-center">
+              {{ T.qualityFormThanksForReview }}
+            </div>
+          </template>
         </div>
-      </div>
-      <div class="button-row text-right">
-        <button
-          class="col-md-4 mr-2 btn btn-primary"
-          type="submit"
-          :disabled="qualitySeal && !tag"
-          @click="slotProps.onSubmit"
-        >
-          {{ T.wordsSend }}
-        </button>
-        <button
-          class="col-md-4 btn btn-secondary"
-          type="button"
-          @click="slotProps.onHide(true)"
-        >
-          {{ T.wordsCancel }}
-        </button>
-      </div>
-    </template>
-  </omegaup-popup>
+      </form>
+    </transition>
+  </omegaup-overlay-popup>
 </template>
 
 <script lang="ts">
 import { Vue, Prop, Component } from 'vue-property-decorator';
-import Popup from './Popup.vue';
+import omegaup_OverlayPopup from '../OverlayPopup.vue';
+import { AvailableViews } from './DemotionPopup.vue';
+import omegaup_RadioSwitch from '../RadioSwitch.vue';
 import T from '../../lang';
 import VueTypeaheadBootstrap from 'vue-typeahead-bootstrap';
 
@@ -115,9 +113,15 @@ import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
 import { faTrash } from '@fortawesome/free-solid-svg-icons';
 library.add(faTrash);
 
+interface ProblemTag {
+  text: string;
+  value: string;
+}
+
 @Component({
   components: {
-    'omegaup-popup': Popup,
+    'omegaup-overlay-popup': omegaup_OverlayPopup,
+    'omegaup-radio-switch': omegaup_RadioSwitch,
     'vue-typeahead-bootstrap': VueTypeaheadBootstrap,
     FontAwesomeIcon,
   },
@@ -126,44 +130,91 @@ export default class ReviewerPopup extends Vue {
   @Prop() allowUserAddTags!: boolean;
   @Prop() levelTags!: string[];
   @Prop() problemLevel!: string;
-  @Prop() publicTags!: string[];
+  @Prop({
+    default: () => [
+      'problemLevelAdvancedCompetitiveProgramming',
+      'problemLevelAdvancedSpecializedTopics',
+      'problemLevelBasicIntroductionToProgramming',
+      'problemLevelBasicKarel',
+      'problemLevelIntermediateAnalysisAndDesignOfAlgorithms',
+      'problemLevelIntermediateDataStructuresAndAlgorithms',
+      'problemLevelIntermediateMathsInProgramming',
+    ],
+  })
+  possibleTags!: string[];
+  @Prop({ default: () => [] }) publicTags!: string[];
   @Prop() selectedPublicTags!: string[];
   @Prop() selectedPrivateTags!: string[];
   @Prop() problemAlias!: string;
   @Prop() problemTitle!: string;
 
+  AvailableViews = AvailableViews;
   T = T;
+  currentView: AvailableViews = AvailableViews.Content;
   qualitySeal = true;
   tag = '';
   publicTagsList = this.selectedPublicTags ?? [];
 
-  PROBLEM_CATEGORIES = [
-    'problemLevelAdvancedCompetitiveProgramming',
-    'problemLevelAdvancedSpecializedTopics',
-    'problemLevelBasicIntroductionToProgramming',
-    'problemLevelBasicKarel',
-    'problemLevelIntermediateAnalysisAndDesignOfAlgorithms',
-    'problemLevelIntermediateDataStructuresAndAlgorithms',
-    'problemLevelIntermediateMathsInProgramming',
-  ];
+  get sortedProblemTags(): ProblemTag[] {
+    return this.possibleTags
+      .map(
+        (x: string): ProblemTag => {
+          return {
+            value: x,
+            text: T[x],
+          };
+        },
+      )
+      .sort((a: ProblemTag, b: ProblemTag): number => {
+        return a.text.localeCompare(b.text, T.lang);
+      });
+  }
 
   addOtherTag(tag: string): void {
     if (!this.publicTagsList.includes(tag)) {
       this.publicTagsList.push(tag);
     }
   }
+
   publicTagsSerializer(tagname: string): string {
     if (Object.prototype.hasOwnProperty.call(T, tagname)) {
       return T[tagname];
     }
     return tagname;
   }
+
   getName(alias: string): string {
     return T[alias];
   }
+
   removeTag(name: string) {
     let pos = this.publicTagsList.indexOf(name);
     this.publicTagsList.splice(pos, 1);
   }
+
+  onHide(): void {
+    this.$emit('dismiss');
+  }
+
+  onSubmit(): void {
+    this.$emit('submit', this.tag, this.qualitySeal);
+    this.currentView = AvailableViews.Thanks;
+    setTimeout(() => this.onHide(), 2000);
+  }
 }
 </script>
+
+<style lang="scss" scoped>
+ul.tag-select {
+  height: 225px;
+  overflow: auto;
+  border: 1px solid #ccc;
+  background: #fff;
+  list-style-type: none;
+}
+.tag-label {
+  width: -webkit-fill-available;
+  margin-bottom: 0;
+  padding-bottom: 0.5rem;
+}
+</style>
